refactor(mokumo-shop): lift settings helpers out of crates/db [#507]
Wave 3 / S3.1a first slice — the only remaining callers of
`mokumo_db::{get_setup_mode, get_shop_name, is_setup_complete}` read
rows from the `settings` table created by the mokumo-shop migrator, so
the helpers move next to `mokumo_shop::db::initialize_database`. This
keeps the mokumo-shop vertical's public DB API coherent and shrinks
`crates/db/src/lib.rs` to just the platform-generic diagnostics that
still need a home in `kikan::db` before the crate can be deleted in
S3.1b.

Addresses CodeRabbit's review comment on #586 that flagged the split
public API. No behavior change — the helpers keep identical signatures
and the unit tests migrate alongside them, plus three new cases for the
previously untested `is_setup_complete` true/false and `get_shop_name`
stored/absent paths.

Callers updated:
- `crates/kikan/src/auth/repo.rs` (login_reset test helper)
- `services/api/src/lib.rs` (setup_status handler + production-mode guard)
- `services/api/tests/bdd_world/auth_steps.rs`
- `crates/db/tests/first_run.rs` and `user_repo.rs`
