+++ crates/kikan/src/auth/repo.rs
@@ -960,7 +960,7 @@ mod tests {
             }
         }
 
-        let is_complete = mokumo_db::is_setup_complete(&db).await.unwrap();
+        let is_complete = mokumo_shop::db::is_setup_complete(&db).await.unwrap();
         assert!(is_complete);
 
         let pool = db.get_sqlite_connection_pool();
